separated python tests into two files and changed comments to make tests clearer
